Add NLATable analysis and use it in PrefixModules (#2695)
Add NLATable analysis and use it in PrefixModules.  This provides a summary of all nla behavior.  This is intended to unify nla handling under one api.  I expect this api to grow as clients get transitioned.  This eliminates the need for breadcrumbs, though we don't deal with that yet (nor does the analysis handle these for now, the users must).